Original Commit
The application has been set up properly, with all of the correct
Gradle Setting enabled. I took the liberty of deleting almost all of
the default OpModes, leaving only the ones pertaining to the color
sensor (we need to test with them), and replaced them with our ‘Example
TeleOp’ and ‘Test Autonomous’ files, both of which have also been
registered in the FtcOpModeRegister file.


Former-commit-id: 360676978e3a71d0651b9b29bcaff0f7e04fc3d8